Move hooks inside the newly created Interface component
